Reduce the sizes of several ints that were twice as large as was required

+++ src/nes/architecture/decoder.rs
@@ -287,7 +287,8 @@ pub struct Decoder {}
 // This is mostly boilerplate to connect the cpu and instructions.
 // This is used to prevent polluting cpu.rs with a thousand lines of boilerplate.
 impl Decoder {
-    pub fn execute(cpu: &mut Cpu, memory: &mut Memory, opcode: u16) {
+    pub fn execute(cpu: &mut Cpu, memory: &mut Memory, opcode: u8) {
+        // Opcodes range from 0x00 to 0xFF, also know as a u8
         match opcode {
             0x00 => {
                 Opcode0x00::execute(cpu, memory);
